fix error message when nothing selected

+++ exercise7/DeletePost.php
@@ -4,29 +4,32 @@
 error_reporting(E_ALL);
 ini_set("display_errors", 1);
 
-$post = $_POST["checkboxvar"];
+if (isset($_POST['checkboxvar']))
+{
+  $post = $_POST["checkboxvar"];
 
-$mysqli = new mysqli("mysql.eecs.ku.edu", "chancepenner2016", "eeRahL7v", "chancepenner2016");
+  $mysqli = new mysqli("mysql.eecs.ku.edu", "chancepenner2016", "eeRahL7v", "chancepenner2016");
 
-/* check connection */
-if ($mysqli->connect_errno) {
-    printf("Connect failed: %s\n", $mysqli->connect_error);
-    exit();
-}
-
-$arrlength = count($post);
-
-// $sql = "SELECT content FROM Posts WHERE author_id='$id'";
-// $result = $mysqli->query($sql);
-
-
-if ($arrlength > 0) {
-  for($i=0;$i<$arrlength;$i++)
-  {
-    $sql = "DELETE FROM Posts WHERE posts_id=$post[$i]";
-    $mysqli->query($sql);
-    echo "Deleted post " . $post[$i] . "<br>";
+  /* check connection */
+  if ($mysqli->connect_errno) {
+      printf("Connect failed: %s\n", $mysqli->connect_error);
+      exit();
   }
+
+  $arrlength = count($post);
+
+  if ($arrlength > 0) {
+    for($i=0;$i<$arrlength;$i++)
+    {
+      $sql = "DELETE FROM Posts WHERE posts_id=$post[$i]";
+      $mysqli->query($sql);
+      echo "Deleted post " . $post[$i] . "<br>";
+    }
+
+  }
+
+  $mysqli->close();
+  /* close connection */
 
 }
 else
@@ -34,7 +37,6 @@ else
   echo "No post selected!";
 }
 
-$mysqli->close();
-/* close connection */
+
 
 ?>
